requirements.txt
Django==5.2.5
pillow>=10.0.0
whitenoise>=6.6.0
gunicorn>=21.2.0
requests>=2.32.0
djangorestframework>=3.15.2
